functions/notificador/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-secret-manager==2.*
